Use lsstsvc1 UID for LFA ingest.
If the underlying Butler repo datastore does not allow writes from the `lsst`
user in the container, which has UID/GID 1000/1000, then ingest cannot occur.
This commit reworks the deployment (without container changes) to allow writing
as a different UID (in this case, the one used for lsstsvc1 at USDF).

Note that this is at the cost of requiring the initContainer to run as root,
which is normally undesirable.